+++ Rebus.AzureStorage.Tests/AzureStorageQueuesTransportMessageExpiration.cs
@@ -1,7 +1,7 @@
 using NUnit.Framework;
 using Rebus.Tests.Contracts.Transports;
 
-namespace Rebus.AzureStorageQueues.Tests
+namespace Rebus.AzureStorage.Tests
 {
     [TestFixture, Category(TestCategory.Azure)]
     public class AzureStorageQueuesTransportMessageExpiration : MessageExpiration<AzureStorageQueuesTransportFactory> { }