+++ CHANGELOG.md
@@ -909,7 +909,7 @@
 * Updated RabbitMQ dependency to 3.6.1
 * Updated Newtonsoft JSON.NET dependency to 8.0.3 across the board
 * Implemented ability for RabbitMQ transport to skip exchange and queue declarations as well as input queue bindings
-
+* Renamed `Rebus.AzureStorageQueues` NuGet package to `Rebus.AzureStorage` because it has integration things for Azure storage
 
 
 [AndreaCuneo]: https://github.com/AndreaCuneo